quality of life updates and new button component
fixed double specification of argument in a header file
forgot to change extension on the input_manager header file(.h -> .hpp)
now in order to render a component YOU NEED to pass the reference to BaseApp, because we need to register some events in the components for example(we're doing it on the first draw)
added a button which has a test texture for now.
the input manager now fully uses event signaling.